Fixes ApiBuilder tests
The ApiBuilder tests now run.

This helped catch a final bug with roles/permissions being rendered for the first time, which was fixed here too.

The decorate test from ApiBuilder is being skipped for now, due to the usage of Select, but will be fixed in another commit later too.